basic initialization process based on etcd (#228)
* first rough etcd init process & renaming node -> service

* introduced service_registry class for decoupling/hiding etcd specifics

* all services now make use of the service_registry

* dropped _service suffix in naming scheme (as everything not in common is a service)

* simplistic waiting for dependencies

* fix compilation errors in test cases

* fix compilation error in testing client

* temporarily disabled metrics exposer (needs refactoring anyhow)

* temporarily disabled read_cache test

* cleanup/more consistent nomenclature

* don't pass boost endpoint but rather pair of ip and port

* improved handling of abbreviation_by_role and role_by_abbreviation maps

* don't use hard-coded /uh/ prefix

* move unregister_service into dtor

* move wait_for_dependency and create_connections into run method for consistency reasons

* switched back to using absolute include paths in entrypoint

* make service registry endpoint URL configurable via program args

* attempt to fix compilation problem in cluster_fixture.h

* attempt to fix compilation problem in testing client

---------

+++ testing/client/main.cpp
@@ -66,7 +66,7 @@ int main (int argc, char* args[]) {
     }
 
     auto ioc = std::make_shared <boost::asio::io_context> ();
-    messenger m (ioc, ps.address, static_cast <int> (ps.port));
+    messenger m (ioc, ps.address, static_cast <std::uint16_t> (ps.port));
 
     LOG_INFO() << "Connected to the server";
